# backend/src/config/openapi.generated.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Signet Workforce ERP API",
    "version": "1.0.0",
    "description": "Production REST API for Signet Workforce ERP. Authenticate via POST /api/auth/login, then use the Bearer token for protected routes."
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Local"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/attendance": {
      "get": {
        "tags": [
          "Attendance"
        ],
        "summary": "List attendance records",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "pageSize",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "employeeId",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "in": "query",
            "name": "siteId",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "in": "query",
            "name": "fromDate",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "in": "query",
            "name": "toDate",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated attendance records"
          }
        }
      }
    },
    "/api/attendance/mark": {
      "post": {
        "tags": [
          "Attendance"
        ],
        "summary": "Mark attendance for an employee",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "employeeId",
                  "attendanceDate",
                  "status"
                ],
                "properties": {
                  "employeeId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "attendanceDate": {
                    "type": "string",
                    "format": "date"
                  },
                  "status": {
                    "type": "integer"
                  },
                  "siteId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "checkInTime": {
                    "type": "string"
                  },
                  "checkOutTime": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Attendance marked; returns id"
          }
        }
      }
    },
...
    "/api/reports/attendance": {
      "get": {
        "tags": [
          "Reports"
        ],
        "summary": "Attendance summary report",
        "parameters": [
          {
            "in": "query",
            "name": "month",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "year",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "fromDate",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "in": "query",
            "name": "toDate",
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Attendance report"
          }
        }
      }
    },
...
  },
...
}